+++ common/config/models_main.go
@@ -4,6 +4,7 @@ type GeneralConfig struct {
 	BindAddress      string `yaml:"bindAddress"`
 	Port             int    `yaml:"port"`
 	LogDirectory     string `yaml:"logDirectory"`
+	LogColors        bool   `yaml:"logColors"`
 	TrustAnyForward  bool   `yaml:"trustAnyForwardedAddress"`
 	UseForwardedHost bool   `yaml:"useForwardedHost"`
 }